bin/requirements.txt
PyGithub==1.55
python-dotenv==0.19.1
PyNaCl==1.4.0
urllib3==1.26.7
